Replaced Chinese implementation
The CSS specification variations did not line up with the variatinos
offered by the chinese_number crate. This new set of implementations not
only supports the variations in the chinese_number crate, but it also
supports the variations the CSS specification utilizes.

All scales are implemented.

The other added bonus is that this crate can now format all systems
without alloc.